feat: formulario inicial com diferenciação entre jogador e treinador
