+++ package.json
@@ -17,6 +17,8 @@
     "@radix-ui/react-toggle-group": "^1.0.4",
     "@radix-ui/react-tooltip": "^1.0.7",
     "@radix-ui/react-visually-hidden": "^1.0.3",
+    "@radix-ui/react-tabs": "1.0.4",
+    "@radix-ui/colors": "3.0.0",
     "axios": "^1.6.8",
     "chart.js": "^4.4.2",
     "react": "^18.2.0",
