+++ src/pages/Register/assets/Register.styled.js
@@ -1,6 +1,6 @@
 import * as LS from '@pages/Login/Login.styles'
 import Colors from "@utils/Colors";
-import styled from 'styled-components'
+import styled, { css } from 'styled-components';
 
 export const Mandatory = styled.span`
    color: ${Colors.red};
@@ -11,7 +11,7 @@ export const StepperWrapper = styled.div`
    top: 13%;
 
    @media (min-width: 1700px) {
-      top: 14%;
+     top: 14%;
    }
 `
 
@@ -34,12 +34,10 @@ export const Text = styled.span`
    top: 28%;
    left: 65%;
    
-
    @media (min-width: 1700px) {
       font-size: 1.5rem;
    }
 `
-
 export const InputLine = styled.div`
    position: relative;
    display: flex;
@@ -47,10 +45,37 @@ export const InputLine = styled.div`
    width: 100%;
    gap: 1rem;
    justify-content: space-between;
-`
+`;
 
-export const InputContainer = styled.div`
-   width: fit-content;
-   display: flex;
-   gap: 1rem;
-`
+export const TabsRoot = styled.div`
+  position: relative;
+  flex-direction: column;
+`;
+
+export const TabsList = styled.div`
+  flex-shrink: 0;
+  display: flex;
+  background: #FFEAE0;
+`;
+
+export const TabsTrigger = styled.button`
+  font-family: inherit;
+  background-color: #131313;
+  height: 45px;
+  flex: 1;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  font-size: 15px;
+  color: #FFEAE0;
+  border: none;
+ 
+  ${props => props.active && `
+    border-bottom: 4px solid #FF7425;
+  `}
+`;
+
+export const TabsContent = styled.div `{
+    flex - grow: 1;
+    padding: 5%;
+}`
