+++ src/pages/Register/index.jsx
@@ -1,65 +1,53 @@
 import { useState } from 'react';
-import * as S from './Register.styled';
+import * as S from '../Register/assets/Register.styled';
 import * as LS from '@pages/Login/Login.styles';
 import Background from '@components/Background/Background';
 import Stepper from '@components/Stepper/Stepper';
-import FormStepOne from './FormStepOne';
-import FormStepTwo from './FormStepTwo';
-import FormStepThree from './FormStepThree';
+import FormStepOne from '../UserRegister/FormStepOne';
+import FormStepTwo from '../UserRegister/TrainerRegister/FormStepTwo';
+import FormStepThree from '../UserRegister/TrainerRegister/FormStepThree';
 import user from '@api/user';
+import team from '@api/team';
 
 export default function Register() {
    const [step, setStep] = useState(1);
    const [userData, setUserData] = useState({
+      name: null,
       email: null,
-      password: null,
-      coach: {
-         firstName: null,
-         lastName: null,
-         birthDate: null,
-         phone: null,
-         picture: null
-      }
+      lastName: null,
+      birthDate: null,
+      phone: null,
+      password: null
    });
-
    const [teamData, setTeamData] = useState({
       code: null,
       name: null,
       category: null,
-      picture: null,
-      local: null,
-      isAmateur: null
-   });
+      logo: null,
+      isAmateur: null,
+   })
 
    function handleFormSubmit(formData) {
       if(step == 1) {
-         userData.coach.firstName = formData.name;
-         userData.coach.lastName = formData.lastName;
-         userData.coach.birthDate = formData.date;
+         userData.name = formData.name;
+         userData.lastName = formData.lastName;
+         userData.birthDate = formData.date;
          setStep(step + 1);
       }
       else if(step == 2) {
          userData.email = formData.email;
+         userData.phone = formData.phone;
          userData.password = formData.password;
-         userData.coach.phone = formData.formattedPhone;
-
-         console.log(userData);
-
-         user.post(userData)
-            .then(response => {
-               console.log(response);
-            })
-            .catch(err => {
-               console.error(err);
-            });
+         user.post(userData);
          setStep(step + 1);
       }
-      else {
+      else if (step == 3){
          teamData.code = formData.teamCode;
          teamData.name = formData.teamName;
          teamData.category = formData.category;
-         teamData.logo = formData.teamPicture;
-         teamData.isAmateur = formData.chkAmateur;
+         teamData.logo = formData.teamLogo;
+          teamData.isAmateur = formData.chkAmateur;
+          team.post(teamData);
       }
    }
