+++ src/pages/UserRegister/AthleteRegister/FormStepTwo.jsx
@@ -1,5 +1,5 @@
 import { useState, useEffect } from 'react';
-import * as S from '../../Register/assets/Register.styled';
+import * as S from './assets/Register.styled';
 import * as LS from '../../Login/Login.styles';
 import Label from '@components/Label/Label';
 import Input from '@components/Input/Input';
@@ -74,8 +74,7 @@ export default function FormStepTwo({onSubmit}) {
          ConfirmPasswordValidation(password, confirmPassword) &&
          BrPhoneValidation(phone)) 
       {
-         let formattedPhone = phone.replace(" ", "").replace("(", "").replace(")", "").replace("-", "");
-         onSubmit({email, formattedPhone, password});
+         onSubmit({email, phone, password});
       }
       else {
          if(!EmailValidation(email)) toast.error('Email inválido.');
