+++ src/pages/UserRegister/TrainerRegister/FormStepThree.jsx
@@ -1,5 +1,5 @@
 import { useState, useEffect } from 'react';
-import * as S from './Register.styled';
+import * as S from '../../Register/assets/Register.styled';
 import * as LS from '../../Login/Login.styles';
 import Label from '@components/Label/Label';
 import Input from '@components/Input/Input';
